requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0
click>=8.1.0
huggingface-hub>=0.20.0
transformers==4.46.3
tokenizers==0.20.3
einops
addict
easydict
PyMuPDF>=1.23.0
python-docx>=1.0.0
Pillow>=10.0.0
python-multipart>=0.0.6
